// src/app/shared/models/RealEstate.model.ts
export class RealEstate {
  id?: string;
  title: string;
  price: number;
  description: string;
  area: number;
  interface?: string;
  bedRooms?: number;
  flats?: number;
  lounges?: number;
  bathRooms?: number;
  streetWidth: number;
  estateAge: number;
  options: [string];
  imagesUrls: any;
  views: number;
  createdAt: Date;
  createdFrom: Date;
  userName: string;
  userImage: string;
  hashId: number;
  location: any;



  // constructor(ralEstate?) {
  //   this.title = ralEstate.title;
  //   this.price = ralEstate.price;
  //   this.description = ralEstate.description;
  //   this.area = ralEstate.area;
  //   this.interface = ralEstate.interface;
  //   this.bedRooms = ralEstate.bedRooms;
  //   this.flats = ralEstate.flats;
  //   this.lounges = ralEstate.lounges;
  //   this.bathRooms = ralEstate.bathRooms;
  //   this.streetWidth = ralEstate.streetWidth;
  //   this.estateAge = ralEstate.estateAge;
  //   this.options = ralEstate.options;
  //   this.imagesUrls = ralEstate.imagesUrls;
  //   this.views = ralEstate.views;
  //   this.createdAt = ralEstate.creadtedAt;
  //   this.createdFrom = ralEstate.date;
  //   this.userName = ralEstate.userName;
  //   this.userImage = ralEstate.UserImage;
  //   this.hashId = ralEstate.hashId;
  //   this.location = ralEstate.location;
  // }
}